Check if a decison should be made
Checks to see if the concentration value is greater than 0.9 or less then 0.1
for longer then 30 seconds. If it has mark robot has having made a decision.

Once a decision is made the belief should not be updated again and the
robot should stay in communication mode to provide positive feed back
for other nodes

Signed-off-by: Andi Gunderson <gunde406@umn.edu>